ui update

+++ app/src/main/java/com/persia/test/data/network/services/persiaatlas/responses/variant/ProductType.kt
@@ -1,7 +1,7 @@
 package com.persia.test.data.network.services.persiaatlas.responses.variant
 
 data class ProductType(
-    val id: Int,
-    val selector_type: Int,
+    val id: Long,
+    val selector_type: Long,
     val title: String
 )
